Organized new Beammap code a bit
